# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UPcar - ride sharing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/scheduled-rides:
    get:
      operationId: getAdminScheduledRides
      tags: [admin]
      summary: All scheduled rides (admin)
      parameters:
        - name: scheduledStatus
          in: query
          schema:
            type: string
            enum: [pending_acceptance, confirmed, driver_declined, cancelled]
        - name: schedulingType
          in: query
          schema:
            type: string
            enum: [public, directed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ride"
  /admin/scheduled-rides/{id}/cancel:
    patch:
      operationId: adminCancelScheduledRide
      tags: [admin]
      summary: Cancel a scheduled ride (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCancelInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/scheduled-rides/{id}/reassign:
    patch:
      operationId: adminReassignScheduledRide
      tags: [admin]
      summary: Reassign a scheduled ride to a different driver (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReassignInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── AUTH ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalRides:
          type: integer
        accountStatus:
          type: string
          enum: [pending, approved, denied]
        createdAt:
          type: string
          format: date-time
        driverProfile:
          $ref: "#/components/schemas/DriverProfile"
...
    DriverProfile:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, denied]
        cpf:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        cnhNumber:
          type: ["string", "null"]
        cnhCategory:
          type: ["string", "null"]
        cnhExpiry:
          type: ["string", "null"]
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleYear:
          type: ["integer", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
          enum: [sedan, suv, hatch, pickup, moto, van, null]
        photoUrl:
          type: ["string", "null"]
        cnhPhotoUrl:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        criminalRecordUrl:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
...
    Ride:
      type: object
      required: [id, passengerId, originAddress, originLat, originLng, destinationAddress, destinationLat, destinationLng, offeredPrice, status, isScheduled, createdAt]
      properties:
        id:
          type: integer
        passengerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        originAddress:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        offeredPrice:
          type: number
        agreedPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, negotiating, accepted, in_progress, completed, cancelled]
        estimatedDistance:
          type: ["number", "null"]
        estimatedDuration:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        isScheduled:
          type: boolean
        scheduledFor:
          type: ["string", "null"]
          format: date-time
        schedulingType:
          type: ["string", "null"]
          enum: [public, directed, null]
        directedToDriverId:
          type: ["integer", "null"]
        scheduledStatus:
          type: ["string", "null"]
          enum: [pending_acceptance, confirmed, driver_declined, cancelled, null]
        scheduledNote:
          type: ["string", "null"]
        passenger:
          $ref: "#/components/schemas/User"
        driver:
          $ref: "#/components/schemas/User"
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        driverLat:
          type: ["number", "null"]
        driverLng:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...
    AdminCancelInput:
      type: object
      properties:
        reason:
          type: ["string", "null"]
    AdminReassignInput:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: integer
...
    Offer:
      type: object
      required: [id, rideId, driverId, price, status, createdAt]
      properties:
        id:
          type: integer
        rideId:
          type: integer
        driverId:
          type: integer
        price:
          type: number
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected]
        driver:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
...