# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * UPcar - ride sharing API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverProfile } from './driverProfile';
import type { UserAccountStatus } from './userAccountStatus';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  rating?: number | null;
  totalRides?: number;
  accountStatus?: UserAccountStatus;
  createdAt: Date;
  driverProfile?: DriverProfile;
}